# openapi.yml
openapi: 3.0.0
info:
  title: School Management System API
  version: 1.0.0
  description: API documentation for School Management System
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.production.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
          example: false
        message:
          type: string
          example: An error occurred
      required:
        - success
        - message
...
    TimetableResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: The timetable data (structure may vary).
        message:
          type: string
          example: Timetable generated successfully
...
security:
  - bearerAuth: []
paths:
...
  /timetable/generate/{classid}:
    post:
      tags:
        - Timetable
      summary: Generate a new timetable for a class
      description: Generates and saves a new timetable schedule for the specified class ID.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classid
          required: true
          description: The ID of the class to generate the timetable for.
          schema:
            type: string
            example: '101'
      responses:
        '200':
          description: Timetable generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableResponse'
        '400':
          description: Invalid input parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error while generating timetable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /timetable/weekly/{classId}/{sectionId}:
    get:
      tags:
        - Timetable
      summary: Get weekly timetable for a class section
      description: Retrieves the weekly timetable schedule for the specified class and section.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: classId
          required: true
          description: The ID of the class.
          schema:
            type: string
            example: '101'
        - in: path
          name: sectionId
          required: true
          description: The ID of the section.
          schema:
            type: string
            example: '1'
      responses:
        '200':
          description: Returns the weekly timetable data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimetableResponse'
        '404':
          description: Timetable not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error while fetching timetable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...